pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "promptmatrix"
version = "0.1.0"
description = "Enterprise-grade prompt governance infrastructure for AI applications."
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "PromptMatrix", email = "hello@promptmatrix.ai"}
]
keywords = ["prompt-engineering", "llm", "governance", "cicd", "ai"]
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy>=2.0.25",
    "alembic>=1.13.1",
    "pydantic-settings>=2.1.0",
    "python-multipart>=0.0.6",
    "httpx>=0.26.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "jinja2>=3.1.3"
]

[project.urls]
Homepage = "https://github.com/PromptMatrix/PromptMatrix"
Documentation = "https://github.com/PromptMatrix/PromptMatrix#readme"

[project.scripts]
pmx = "pmx:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.ruff]
line-length = 88
target-version = "py311"
